change room version to 2.3.0-aplha02 to support paging 3 - change photo id to Int

+++ app/src/main/java/dev/leonlatsch/photok/model/database/dao/PhotoDao.kt
@@ -18,5 +18,5 @@ interface PhotoDao {
     suspend fun delete(photo: Photo)
 
     @Query("SELECT * FROM photo ORDER BY importedAt DESC")
-    fun getAllPagedSortedByImportedAt(): PagingSource<Long, Photo>
+    fun getAllPagedSortedByImportedAt(): PagingSource<Int, Photo>
 }
